#14 Only remove link if it exists

+++ extensions/user_template/Classes/Composer/InstallerScripts.php
@@ -122,7 +122,9 @@ class InstallerScripts implements InstallerScriptsRegistration, InstallerScript
         $target = $sourcePath . DIRECTORY_SEPARATOR . $file;
         $target = $this->getRelativePath($link, $target);
         echo "$target => $link\n";
-        unlink($link);
+        if(file_exists($link)) {
+            unlink($link);
+        }
         symlink($target, $link);
     }
 }
